Update movie platform with download system and payment integration

+++ frontend_movie/src/admin/Config/authApi.js
@@ -12,33 +12,35 @@ export const getAllUser = () => {
     return httpClient.get(`/api/users/`);
 };
 
+
+//movie
 export const getAllMovies = (page = 0) => {
-    return httpClient.get(`/api/movie/?pageNumber=${page}`);
+    return httpClient.get(`/api/admin/movie/?pageNumber=${page}`);
 };
 
 export const addMovie = (data)=>{
-  return httpClient.post(`/api/movie/`,data)
+  return httpClient.post(`/api/admin/movie/`,data)
 }
 
 export const deleteMovies = (movieId) => {
-    return httpClient.delete(`/api/movie/${movieId}`);
+    return httpClient.delete(`/api/admin/movie/${movieId}`);
 };
 
 export const searchMovie = (movieName) => {
-    return httpClient.get(`/api/movie/search/${movieName}`);
+    return httpClient.get(`/api/admin/movie/search/${movieName}`);
 };
 
 export const updateMovie = (movieId, data) => {
-  return httpClient.put(`/api/movie/${movieId}`, data);
+  return httpClient.put(`/api/admin/movie/${movieId}`, data);
 };
 
 export const getMovieById = (movieId) => {
-    return httpClient.get(`/api/movie/${movieId}`)
+    return httpClient.get(`/api/admin/movie/${movieId}`)
 }
 
 // SCREENSHOTS
 export const getScreenshots = (movieId) => {
-    return httpClient.get(`/api/movie/${movieId}/screenshots`);
+    return httpClient.get(`/api/admin/movie/${movieId}/screenshots`);
   };
   
 
@@ -48,7 +50,7 @@ export const uploadScreenshots = (movieId, files) => {
   files.forEach(file => formData.append("images", file));
 
   return httpClient.post(
-    `/api/movie/${movieId}/screenshots`,
+    `/api/admin/movie/${movieId}/screenshots`,
     formData,
     {
       headers: {
@@ -62,7 +64,7 @@ export const updateThumbnail = (movieId, file) => {
   const formData = new FormData();
   formData.append("image", file);
 
-  return httpClient.put(`/api/movie/${movieId}/thumbnail`, formData, {
+  return httpClient.put(`/api/admin/movie/${movieId}/thumbnail`, formData, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -73,7 +75,7 @@ export const updateScreenShot = (selectedSS, movieId, file) => {
   const formData = new FormData();
   formData.append("image", file);
 
-  return httpClient.put(`/api/movie/${movieId}/screenshot/${selectedSS}`, formData, {
+  return httpClient.put(`/api/admin/movie/${movieId}/screenshot/${selectedSS}`, formData, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -81,23 +83,23 @@ export const updateScreenShot = (selectedSS, movieId, file) => {
 };
 
 export const deleteScreenshot = (movieId, screenshotId) => {
-  return httpClient.delete(`/api/movie/${movieId}/screenshot/${screenshotId}`);
+  return httpClient.delete(`/api/admin/movie/${movieId}/screenshot/${screenshotId}`);
 };
 
 
 // QUALITY
 export const getQualities = (movieId) => {
-  return httpClient.get(`/api/movie/${movieId}/quality`);
+  return httpClient.get(`/api/admin/movie/${movieId}/quality`);
 };
 
 export const addQuality = (movieId, data) => {
-  return httpClient.post(`/api/movie/${movieId}/quality`, data);
+  return httpClient.post(`/api/admin/movie/${movieId}/quality`, data);
 };
 
 export const updateQuality = (qualityId, data) => {
-  return httpClient.put(`/api/movie/quality/${qualityId}`, data);
+  return httpClient.put(`/api/admin/movie/quality/${qualityId}`, data);
 };
 
 export const deleteQuality = (qualityId) => {
-  return httpClient.delete(`/api/movie/quality/${qualityId}`);
+  return httpClient.delete(`/api/admin/movie/quality/${qualityId}`);
 };
